refactor(api): split users routes integration test per use case
- Extract shared app lifecycle into abstract BaseUsersIntegrationTest
- Name test classes after use cases instead http verbs:
  GetAllUsers, CreateUser, UpdateUserById, GetUserById, DeleteUserById
- Group cross-route HTTP contract tests in UsersContractIntegrationTest
- Stop Application with a shutdown hook
