+++ backend/api/src/test/java/com/anibalxyz/features/users/application/CreateUserTest.java
@@ -22,6 +22,7 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 @ExtendWith(MockitoExtension.class)
+@DisplayName("Tests for CreateUser service")
 public class CreateUserTest {
   @Mock private UserRepository userRepository;
