+++ backend/api/src/test/java/com/anibalxyz/features/users/application/DeleteUserByIdTest.java
@@ -17,6 +17,7 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 @ExtendWith(MockitoExtension.class)
+@DisplayName("Tests for DeleteUserById service")
 public class DeleteUserByIdTest {
   @Mock private UserRepository userRepository;
