+++ backend/api/src/test/java/com/anibalxyz/features/users/application/GetAllUsersTest.java
@@ -17,6 +17,7 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 @ExtendWith(MockitoExtension.class)
+@DisplayName("Tests for GetAllUsers service")
 public class GetAllUsersTest {
   @Mock private UserRepository userRepository;
